requirements.txt
grpcio
grpcio-tools
redis
pyyaml
hashlib
lru-dict
psutil
pysyncobj